refactor [#446] 클래스 명 변경
- SearchTermController -> SearchController
