+++ src/main/java/org/sopt/confeti/api/search/controller/SearchController.java
@@ -1,11 +1,11 @@
-package org.sopt.confeti.api.search_term.controller;
+package org.sopt.confeti.api.search.controller;
 
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
 import lombok.RequiredArgsConstructor;
-import org.sopt.confeti.api.search_term.dto.response.PopularTermsResponse;
-import org.sopt.confeti.api.search_term.facade.SearchTermFacade;
-import org.sopt.confeti.api.search_term.facade.dto.response.PopularTermsDTO;
+import org.sopt.confeti.api.search.dto.response.PopularTermsResponse;
+import org.sopt.confeti.api.search.facade.SearchFacade;
+import org.sopt.confeti.api.search.facade.dto.response.PopularTermsDTO;
 import org.sopt.confeti.domain.user.constant.Role;
 import org.sopt.confeti.global.annotation.Permission;
 import org.sopt.confeti.global.annotation.UserId;
@@ -20,18 +20,18 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/search/terms")
-public class SearchTermController {
+@RequestMapping("/search")
+public class SearchController {
 
-    private final SearchTermFacade searchTermFacade;
+    private final SearchFacade searchFacade;
 
     @Permission(role = {Role.GENERAL})
-    @GetMapping("/popular")
+    @GetMapping("/terms/popular")
     public ResponseEntity<BaseResponse<?>> getPopularSearchTerms(
             @UserId(require = false) Long userId,
             @RequestParam(required = false, defaultValue = "10") @Min(1) @Max(20) Integer limit
     ) {
-        PopularTermsDTO terms = searchTermFacade.getPopularSearchTerms(limit);
+        PopularTermsDTO terms = searchFacade.getPopularSearchTerms(limit);
         return ApiResponseUtil.success(SuccessMessage.SUCCESS, PopularTermsResponse.from(terms));
     }
 }
