+++ src/main/java/org/sopt/confeti/api/search/dto/response/PopularTermResponse.java
@@ -1,6 +1,6 @@
-package org.sopt.confeti.api.search_term.dto.response;
+package org.sopt.confeti.api.search.dto.response;
 
-import org.sopt.confeti.api.search_term.facade.dto.response.PopularTermDTO;
+import org.sopt.confeti.api.search.facade.dto.response.PopularTermDTO;
 
 public record PopularTermResponse(
         int rank,
