+++ src/main/java/org/sopt/confeti/api/search/facade/SearchFacade.java
@@ -1,13 +1,13 @@
-package org.sopt.confeti.api.search_term.facade;
+package org.sopt.confeti.api.search.facade;
 
 import lombok.RequiredArgsConstructor;
-import org.sopt.confeti.api.search_term.facade.dto.response.PopularTermsDTO;
+import org.sopt.confeti.api.search.facade.dto.response.PopularTermsDTO;
 import org.sopt.confeti.domain.elastic_search.application.SearchTermService;
 import org.sopt.confeti.global.annotation.Facade;
 
 @Facade
 @RequiredArgsConstructor
-public class SearchTermFacade {
+public class SearchFacade {
 
     private final SearchTermService searchTermService;
