+++ src/main/java/org/sopt/confeti/api/search/facade/dto/response/PopularTermsDTO.java
@@ -1,4 +1,4 @@
-package org.sopt.confeti.api.search_term.facade.dto.response;
+package org.sopt.confeti.api.search.facade.dto.response;
 
 import java.util.List;
 import org.sopt.confeti.domain.elastic_search.application.dto.response.PopularTermResult;
